+++ examples/copyConfigForTestsOnly/config/log.js
@@ -4,4 +4,4 @@ module.exports = {
   writeBuffer:    64*1024,
   applicationLog: false,
   serverLog:      true,
-};
+}