+++ examples/copyContentToProjectFolder/applications/example/config/sandbox.js
@@ -47,4 +47,4 @@ module.exports = {
     'impress.state',
   ]
 
-};
+}