+++ examples/impressCloudController/instanceC0/config/servers.js
@@ -8,18 +8,5 @@ module.exports = {
     port:     80,
     nagle:    true, // Nagle algorithm, default true, set to false for latency optimization
     slowTime: '1s'
-  },
-  //ssl: {
-  //  protocol:  'https',
-  //  address:   '127.0.0.1',
-  //  port:      443,
-  //  key:       'example.key',
-  //  cert:      'example.cer'
-  //},
-  //static: {
-  //  protocol:  'http',
-  //  address:   '127.0.0.1',
-  //  port:      8080,
-  //  slowTime:  1000
-  //}
-};
+  }
+}