+++ examples/impressCloudController/instanceC2/applications/example/config/application.js
@@ -4,4 +4,4 @@ module.exports = {
   slowTime:    '1s',
   // preload:     true, // preload application to memory at startup (default: false)
   // allowOrigin: '*' // set HTTP header Access-Control-Allow-Origin (default: not set)
-};
+}