+++ examples/impressCloudController/instanceC2/applications/example/config/files.js
@@ -1,8 +1,7 @@
 // Static files configuration
 
 module.exports = {
-  minify: false,  // minify JavaScript files
   static: [ '*/css/*', '*/images/*', '*/js/*', '*/favicon.ico', '*/favicon.png' ],  // Masks for static files
   cacheSize: '50mb',  // memory cache size
   cacheMaxFileSize: '10mb',  // max file size to cache
-};
+}