+++ examples/impressCloudController/instanceC2/config/log.js
@@ -7,4 +7,4 @@ module.exports = {
   applicationLog: false,   // Write log to application folder
   serverLog:      true,    // Write log to server global folder
   stdout:         ['error', 'debug'] // output log files to stdout
-};
+}